Simple changes to the top-level requirements
 - Added keeping track of position at all times to manual mode as well
 - Added monitoring for Mode-Switch messages to manual mode as well

This commit is to make the top-level requirements a bit more complete,
adding things that I think were missed in my opinion. The position
tracking I added because it'll make transitioning back into autonomous
easier (which the ability to do is in the requirements already). The
mode-switch monitoring in manual is mainly just to make it consistent
between autonomous and manual mode.